requirements.txt
requests>=2.31.0
pytos2-ce>=3.0.0
xmltodict>=0.13.0
python-dotenv>=1.0.1

